+++ lib/ui/page/home/page/chat/widget/swipeable_status.dart
@@ -84,7 +84,15 @@ class SwipeableStatus extends StatelessWidget {
           ? Alignment.bottomRight
           : Alignment.centerRight,
       children: [
-        translate ? _animatedBuilder(child, translated: false) : child,
+        translate
+            ? _animatedBuilder(
+                Padding(
+                  padding: const EdgeInsets.only(left: width),
+                  child: child,
+                ),
+                translated: false,
+              )
+            : child,
         _animatedBuilder(
           Padding(
             padding: padding,