+++ CHANGELOG.md
@@ -41,7 +41,7 @@ All user visible changes to this project will be documented in this file. This p
     - Home page:
         - Redesigned chats tab. ([#142])
         - Redesigned introduction and logout modals. ([#249])
-        - Redesigned menu tab. ([#244], [#243])
+        - Redesigned menu tab. ([#313], [#244], [#243])
     - Media panel:
         - Video resizing when dragged. ([#191], [#190])
         - Redesigned mobile interface. ([#287], [#246])
@@ -56,6 +56,7 @@ All user visible changes to this project will be documented in this file. This p
     - Profile page:
         - Redesigned profile page. ([#244], [#257], [#243])
     - Redesigned desktop context menu. ([#245])
+    - Redesigned mobile context menu. ([#305], [#295])
     - Auth page:
         - Redesigned login modal. ([#249])
     - User page:
@@ -142,8 +143,11 @@ All user visible changes to this project will be documented in this file. This p
 [#288]: /../../pull/288
 [#292]: /../../issues/292
 [#293]: /../../pull/293
+[#295]: /../../issues/295
 [#300]: /../../pull/300
+[#305]: /../../pull/305
 [#310]: /../../pull/310
+[#313]: /../../pull/313